// src/sanity/sanity.types.ts
export type LAYOUT_QUERY_RESULT = {
  info: {
    name: string | null;
    managingDirector: string | null;
    orgNumber: string | null;
    address1: string | null;
    address2: string | null;
    zipCode: string | null;
    city: string | null;
    country: string | null;
    email: string | null;
    phone: string | null;
    googleMapsLatitude: number | null;
    googleMapsLongitude: number | null;
    socialMedias: Array<{
      type: "Facebook" | "GitHub" | "LinkedIn" | "Twitter" | null;
      url: string | null;
    }> | null;
  } | null;
};